docker/kafka-producer/requirements.txt
kafka-python>=1.0